making top bar persistent across all pages of the app
