+++ healthy_app/lib/screens/home/home.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:healthy_app/services/auth.dart';
 import 'package:healthy_app/services/database.dart';
+import 'package:healthy_app/shared/ConstantVars.dart';
 import 'package:provider/provider.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:healthy_app/screens/home/userSettings_list.dart';
-import 'package:flutter_icons/flutter_icons.dart';
 import '../foodDiary.dart';
 import '../activityDiary.dart';
 import '../medicationTracker.dart';
@@ -48,6 +47,32 @@ class _HomeState extends State<Home> {
     return StreamProvider<QuerySnapshot>.value(
       value: DatabaseService().userSettings,
       child: Scaffold(
+        appBar: AppBar(
+          leading: GestureDetector(
+            onTap: () {
+              /* Write listener code here */
+              print("Calendar View Selected");
+            },
+            child: Icon(
+              Icons.calendar_today_outlined,
+            ),
+          ),
+          title: new Text(getCurrentDate()),
+          centerTitle: true,
+          actions: <Widget>[
+            PopupMenuButton<String>(
+              onSelected: choiceAction,
+              itemBuilder: (BuildContext context){
+                return ConstantVars.choices.map((String choice){
+                  return PopupMenuItem<String>(
+                    value: choice,
+                    child: Text(choice),
+                  );
+                })
+                    .toList();
+              }
+              ,)]
+          ,),
         backgroundColor: Colors.white,
         body: PageView(
           controller: _pageController,
@@ -84,5 +109,23 @@ class _HomeState extends State<Home> {
     ),
     ),
     );
+  }
+  void choiceAction(String choice){
+    if(choice == ConstantVars.Settings){
+      print('Settings');
+    }
+    else if(choice == ConstantVars.Subscribe){
+      print('Subscribe');
+    }
+    else if(choice == ConstantVars.SignOut){
+      _auth.signOut();
+      print('SignOut');
+    }
+  }
+  String getCurrentDate(){
+    var date = new DateTime.now().toString();
+    var dateParse = DateTime.parse(date);
+    var formattedDate = "${dateParse.day}/${dateParse.month}/${dateParse.year}";
+    return formattedDate;
   }
 }
